% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtimetc.R
\name{addtimetc}
\alias{addtimetc}
\title{Add time to your timecard}
\usage{
addtimetc(
  client = "sdi",
  pn = "admin",
  description = "",
  started = NULL,
  finished = NULL,
  notes = "",
  date = Sys.Date(),
  existing_timecard = timecard
)
}
\arguments{
\item{client}{Client initials ex. 1pw, gsk, ccc}

\item{pn}{Project Name}

\item{description}{Description of the project}

\item{started}{The time that the work block started. Must be a character string. ex. '08:00'}

\item{finished}{The time that the work block is finished. Must be a character string. ex. '14:30'}

\item{notes}{Add notes for each project if needed}

\item{date}{The date the event took place. Defaults to the current date.}

\item{existing_timecard}{The timecard object that will be used to record the new project activity. Does not need to exist}
}
\description{
This is used to add time to the timecard file and updates the timecard csv for the day.
}
