\name{endday}
\alias{endday}
\title{Generate a csv file of your timecard for the day}
\usage{
endday()
}
\description{
Collects the 'timecard' dataframe in the global environment and converts it to a 'csv' file format. The
file name will include the date of the day the function is run and 'timecard'. Example of the output would look something like this: '2020-02-20_timecard.csv'.
}
\examples{
endday()
}

