% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr-package.r
\docType{package}
\name{gcalendr}
\alias{gcalendr}
\alias{gcalendr-package}
\title{gcalendr}
\description{
You can read Google calendar data using the  Google Calendar API
}
\details{
Access your calendar:

* [calendar_list()]
* [calendar_events()]

Authenticate

* [calendar_auth()]
* [calendar_token()]
}
