% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_todays_events.R
\name{gcaltoday}
\alias{gcaltoday}
\title{Get list of the events scheduled for today (or any day).}
\usage{
gcaltoday(today = Sys.Date(), id = gargle::gargle_oauth_email())
}
\arguments{
\item{token}{token obtained from [calendar_token()]}
}
\value{
Tibble with columns for .
}
\description{
Get list of the events scheduled for today (or any day).
}
\seealso{
Other calendar functions: 
\code{\link{calendar_events}()},
\code{\link{calendar_list}()}
}
\concept{calendar functions}
