% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoretimecard.R
\name{restortc}
\alias{restortc}
\title{Restore the timecard}
\usage{
restortc(filepath = NULL)
}
\arguments{
\item{filepath}{The file path of the most recent timecard csv file}
}
\description{
If the timecard object has been deleted and you want to continue generating the timecard
records.
}
