% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summarize the timecard by date and client}
\usage{
summary(timecard = timecard, flag_date = Sys.Date())
}
\arguments{
\item{timecard}{adds the timecard dataframe to the function}

\item{flag_date}{For current week leave as default. A different date can be added if needed.}
}
\description{
Summarizes the results of your hard work by day for the current week
}
