% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tca_followup.R
\name{tca_followup}
\alias{tca_followup}
\title{Set followup and next step items.}
\usage{
tca_followup(
  how = "email",
  who = "Neha",
  when = "today",
  what = "something very important",
  todo = todos
)
}
\arguments{
\item{how}{What action are you needing to take}

\item{who}{Who are you going to do this for/to}

\item{when}{When will this need to be completed}

\item{what}{What are the details that need to be accomplished}

\item{todos}{This is just pulling in the existing 'todos' dataframe if it exists}
}
\value{
dataframe of todo items
}
\description{
Set followup and next step items.
}
\concept{project management functions}
