% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tca_summary.R
\name{tca_summary}
\alias{tca_summary}
\title{Summarize your timecard by date and client}
\usage{
tca_summary(
  timecard = timecard,
  date_filter = floor_date(as.Date(Sys.Date(), "\%m/\%d/\%Y"), unit = "week")
)
}
\arguments{
\item{timecard}{adds the timecard dataframe to the function}

\item{date_filter}{This helps filter the results to today's date or change it.}
}
\description{
Summarizes the group by day
}
\details{
This summarizes the results of you hard work by day.
}
