% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tca_followup.R
\name{unfinished}
\alias{unfinished}
\title{Summarise the unfinished tasks}
\usage{
unfinished(lookfrom = Sys.Date() - 1, status = "unfinished")
}
\arguments{
\item{lookfrom}{The date to look for unfinished tasks}

\item{status}{The status of the event}
}
\description{
Summarise the unfinished tasks
}
