% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endday.R
\name{endday}
\alias{endday}
\title{End the day and get a csv file}
\usage{
endday()
}
\description{
Use this as often as you like to make sure the global
variables are not erased during your day. The file should be overwritten
the next time you used the function on the same day.
}
\details{
When your day is done and you want to keep a copy of the day's
times for a record. Call this function and it will produce a CSV
file on your home directory.  It will include the date in the filename.
}
