% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starttimer.R
\name{starttimer}
\alias{starttimer}
\title{Start the timer}
\usage{
starttimer(pn, starttime = Sys.time(), started = NA)
}
\arguments{
\item{pn}{Project Name}

\item{starttime}{The time that the project started. If you started
earlier and want to catch up add}

\item{started}{The time delay in decimal hour when you started the project.
For example use '.50' to start the project 30 minutes ago.}
}
\description{
Use a project name that you can reference.
}
\details{
This is used when you are starting a work chunk on a project. It
creates a global variable, 'ps', that is used to define the start
time of a project and helps build the 'timecard' ouput when the done()
function is run.
}
