% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytc.R
\name{summarytc}
\alias{summarytc}
\title{Summarize the timecard by date and client}
\usage{
summarytc(timecardobject = timecard, flag_date = Sys.Date())
}
\arguments{
\item{timecardobject}{adds the timecard dataframe to the function}

\item{flag_date}{For current week leave as default. A different date can be added if needed.}
}
\value{
The summary table showing the current week by default but the argument 'flag_date' can
be utilized to show any other date that has been recorded in the timecard object.
}
\description{
Summarizes the results of your hard work by day for the current week
}
